Decouple DDG Generation and Dot Serialization
